renamed assigner.js to assignerLogic.js
